pyproject.toml
[project]
name = "rag-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "dotenv>=0.9.9",
    "nltk>=3.9.2",
    "numpy>=2.3.5",
    "ollama>=0.6.1",
    "pillow>=12.0.0",
    "sentence-transformers>=5.1.2",
]
